Python_Django/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
APScheduler==3.9.1.post1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
axios==0.1.0
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.2.0
cssutils==2.6.0
Django==4.1.4
django-ninja==0.20.0
django-ratelimit==4.0.0
django-seed==0.3.1
django-simple-tags==0.6.0
django-storages==1.13.2
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.3.0
drf-yasg==1.21.4
email-validator==1.3.1
Faker==15.3.4
frozenlist==1.3.3
future==0.18.2
geoip2==4.6.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
maxminddb==2.2.0
multidict==6.0.4
mysqlclient==2.1.1
oauth2client==4.1.3
outcome==1.2.0
packaging==23.0
pefile==2022.5.30
premailer==3.10.0
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
ratelimit==2.2.1
regex==2022.10.31
requests==2.28.2
rich==13.1.0
rsa==4.9
ruamel.yaml==0.17.21
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
toposort==1.7
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
wsproto==1.2.0
yagmail==0.15.293
yarl==1.8.2
